Moved truncate to its own file
